+++ src/utils/typings.test-d.ts
@@ -2,6 +2,7 @@ import {
   PartialCustom,
   type PARTIAL,
   type TransformPrimitiveObject,
+  type inferT,
 } from './typings';
 
 describe('#01 => CustomPartial', () => {
@@ -33,23 +34,23 @@ describe('#01 => CustomPartial', () => {
 
   it('#04 => should transform with "undefined" as string', () => {
     type TTT4 = TransformPrimitiveObject<{
-      [PARTIAL]: 'undefined';
       age: 'number';
+      [PARTIAL]: 'undefined';
     }>;
 
-    expectTypeOf<TTT4>().toExtend<{
+    expectTypeOf<TTT4>().branded.toEqualTypeOf<{
       age: number;
-      '$$app-ts => partial$$': undefined;
+      '$$app-ts => partial$$'?: undefined;
     }>();
   });
 
   it('#05 => should transform with "undefined" as number', () => {
-    type TTT5 = TransformPrimitiveObject<{
+    type TTT5 = inferT<{
       [PARTIAL]: undefined;
       age: 'number';
     }>;
 
-    expectTypeOf<TTT5>().toExtend<{
+    expectTypeOf<TTT5>().toEqualTypeOf<{
       age?: number | undefined;
     }>();
   });